sensor_fusion: use f_t to avoid casts now that rc_tracker.h is the external interface

+++ corvis/src/filter/sensor_fusion.cpp
@@ -107,14 +107,14 @@ std::vector<sensor_fusion::feature_point> sensor_fusion::get_features() const
         if(i->is_valid()) {
             feature_point p;
             p.id = i->id;
-            p.x = (float)i->current[0];
-            p.y = (float)i->current[1];
-            p.original_depth = (float)i->v.depth();
-            p.stdev = (float)i->v.stdev_meters(sqrt(i->variance()));
+            p.x = i->current[0];
+            p.y = i->current[1];
+            p.original_depth = i->v.depth();
+            p.stdev = i->v.stdev_meters(sqrt(i->variance()));
             v4 ext_pos = filter_to_external_position(i->world);
-            p.worldx = (float)ext_pos[0];
-            p.worldy = (float)ext_pos[1];
-            p.worldz = (float)ext_pos[2];
+            p.worldx = ext_pos[0];
+            p.worldy = ext_pos[1];
+            p.worldz = ext_pos[2];
             p.initialized = i->is_initialized();
             features.push_back(p);
         }
@@ -129,14 +129,14 @@ void sensor_fusion::update_data(image_gray8 &&image)
     //perform these operations synchronously in the calling (filter) thread
     d->total_path_m = sfm.s.total_distance;
     camera_parameters cp;
-    cp.fx = (float)sfm.s.focal_length.v * sfm.s.image_height;
-    cp.fy = (float)sfm.s.focal_length.v * sfm.s.image_height;
-    cp.cx = (float)sfm.s.center_x.v * sfm.s.image_height + sfm.s.image_width / 2. - .5;
-    cp.cy = (float)sfm.s.center_y.v * sfm.s.image_height + sfm.s.image_height / 2. - .5;
+    cp.fx = sfm.s.focal_length.v * sfm.s.image_height;
+    cp.fy = sfm.s.focal_length.v * sfm.s.image_height;
+    cp.cx = sfm.s.center_x.v * sfm.s.image_height + sfm.s.image_width / 2. - .5;
+    cp.cy = sfm.s.center_y.v * sfm.s.image_height + sfm.s.image_height / 2. - .5;
     cp.skew = 0;
-    cp.k1 = (float)sfm.s.k1.v;
-    cp.k2 = (float)sfm.s.k2.v;
-    cp.k3 = (float)sfm.s.k3.v;
+    cp.k1 = sfm.s.k1.v;
+    cp.k2 = sfm.s.k2.v;
+    cp.k3 = sfm.s.k3.v;
     d->camera_intrinsics = cp;
 #ifdef ENABLE_QR
     if(sfm.qr.valid)
